# PERF: Improve performance for pyarrow to_numpy with na and na_value